test: Move custom matchers to tests/CustomMatchers/
