+++ tests/Query/Filter/DescriptionField.test.ts
@@ -3,7 +3,7 @@ import { getSettings, updateSettings } from '../../../src/config/Settings';
 import { testTaskFilter } from '../../TestingTools/FilterTestHelpers';
 import { fromLine } from '../../TestHelpers';
 import type { FilterOrErrorMessage } from '../../../src/Query/Filter/Filter';
-import { toMatchTaskWithDescription } from './CustomMatchersForFilters';
+import { toMatchTaskWithDescription } from '../../CustomMatchers/CustomMatchersForFilters';
 
 function testDescriptionFilter(
     filter: FilterOrErrorMessage,
